feat: lone values will now be passed to mergeOthers rather than just returned
Previously if a value had nothing to be merged with, it would just be used as the result. Now it is
instead passed to the mergeOthers function so that the result can be customized if wanted

BREAKING CHANGE: The `mergeOthers` function will now be passed lone values, make sure your custom
version can handle these cases.